# CUDA: mul_mat_q RDNA2 tunings