This PR adds mul_mat_q tunings for RDNA 2. They are enabled for all AMD GPUs if HIP is used because I expect AMD architectures to be more similar to each other than RDNA 2 is to NVIDIA - bug please do check whether these tunings work better than master if you have an AMD GPU other than an RX 6800. These are the numbers that I'm getting:

| Model                        | GPU                |   t/s master |   t/s PR |   Speedup |
|------------------------------|--------------------|--------------|----------|-----------|
| LLaMA 7B mostly Q4_0         | AMD Radeon RX 6800 |      1055.67 |  1209.18 |      1.15 |
| LLaMA 7B mostly Q4_1         | AMD Radeon RX 6800 |       989.55 |  1126.98 |      1.14 |
| LLaMA 7B mostly Q5_0         | AMD Radeon RX 6800 |       963.39 |  1063.83 |      1.10 |
| LLaMA 7B mostly Q5_1         | AMD Radeon RX 6800 |       966.79 |  1065.84 |      1.10 |
| LLaMA 7B mostly Q8_0         | AMD Radeon RX 6800 |      1002.66 |  1130.16 |      1.13 |
| LLaMA 7B mostly Q2_K         | AMD Radeon RX 6800 |       603.98 |   641.65 |      1.06 |
| LLaMA 7B mostly Q3_K - Small | AMD Radeon RX 6800 |       584.14 |   627.74 |      1.07 |
| LLaMA 7B mostly Q4_K - Small | AMD Radeon RX 6800 |       877.90 |  1021.37 |      1.16 |
| LLaMA 7B mostly Q5_K - Small | AMD Radeon RX 6800 |       785.44 |   996.13 |      1.27 |
| LLaMA 7B mostly Q6_K         | AMD Radeon RX 6800 |       819.04 |   920.44 |      1.12 |

As part of this PR I changed the logic regarding compute capability on AMD: for AMD GPUs the compute capability is offset by 1000000 to separate it from the values that NVIDIA GPUs use (e.g. 860 for compute capability 8.6). Changing `CC_TURING` is now no longer necessary.