@ardfork 2023-08-30
With a 6700 XT using gfx1030 override on the models that I have and that I can fully offload:
Model | GPU | t/s master | t/s PR | Speedup
-- | -- | -- | -- | --
LLaMA v2 13B mostly Q4_K | AMD Radeon RX 6700 XT | 369.65 | 433.72 | 1.17
LLaMA v2 13B mostly Q6_K | AMD Radeon RX 6700 XT | 329.90 | 387.64 | 1.18