@Engininja2 2023-08-31
This is with a 5700XT:
Model | test | t/s master | t/s PR | Speedup
-- | -- | -- | -- | --
LLaMA v2 7B mostly Q8_0 | pp 512 | 300.11 | 290.13 | 0.97
LLaMA v2 7B mostly Q6_K | pp 512 | 251.39 | 218.86 | 0.87
LLaMA v2 7B mostly Q5_K - Small | pp 512 | 262.43 | 215.69 | 0.82
LLaMA v2 7B mostly Q5_1 | pp 512 | 290.14 | 275.05 | 0.95
LLaMA v2 7B mostly Q5_0 | pp 512 | 284.69 | 266.82 | 0.94
LLaMA v2 7B mostly Q4_K - Small | pp 512 | 262.99 | 221.83 | 0.84
LLaMA v2 7B mostly Q4_1 | pp 512 | 279.07 | 284.71 | 1.02
LLaMA v2 7B mostly Q4_0 | pp 512 | 303.25 | 294.39 | 0.97
LLaMA v2 7B mostly Q3_K - Small | pp 512 | 199.15 | 190.72 | 0.96
LLaMA v2 7B mostly Q2_K | pp 512 | 191.99 | 175.93 | 0.92