@KerfuffleV2 2023-08-31
I tested on my 6600 RX. TL;DR is token generation isn't really affected, slight speed up for prompt processing. I tested with a 3B `Q8_0` model and a 7B `Q4_K_M` model vs `master`, `master` with `LLAMA_CUDA_DMMV_X=64 LLAMA_CUDA_MMV_Y=8 LLAMA_CUDA_KQUANTS_ITER=1` and this pull (I didn't try setting those overrides on this pull).

Sorry this isn't very scientific but I did run each test several times to make sure the results were stable.

<details>

<summary>
Click to expand
</summary>

## 3B `q8_0`

```plaintext
master
perplexity: 0.56 seconds per pass - ETA 5.72 minutes
llama_print_timings:      sample time =    34.40 ms /  1000 runs   (    0.03 ms per token, 29071.46 tokens per second)
llama_print_timings: prompt eval time =  1290.85 ms /   853 tokens (    1.51 ms per token,   660.80 tokens per second)
llama_print_timings:        eval time = 21112.29 ms /   990 runs   (   21.33 ms per token,    46.89 tokens per second)


master `LLAMA_CUDA_DMMV_X=64 LLAMA_CUDA_MMV_Y=8 LLAMA_CUDA_KQUANTS_ITER=1`
perplexity: 0.54 seconds per pass - ETA 5.55 minutes
llama_print_timings:      sample time =    35.42 ms /  1000 runs   (    0.04 ms per token, 28230.25 tokens per second)
llama_print_timings: prompt eval time =  1283.21 ms /   853 tokens (    1.50 ms per token,   664.74 tokens per second)
llama_print_timings:        eval time = 21166.67 ms /   990 runs   (   21.38 ms per token,    46.77 tokens per second)

pull:
perplexity: 0.50 seconds per pass - ETA 5.08 minutes
llama_print_timings:        load time =   924.38 ms
llama_print_timings:      sample time =    29.48 ms /  1000 runs   (    0.03 ms per token, 33915.55 tokens per second)
llama_print_timings: prompt eval time =  1174.27 ms /   853 tokens (    1.38 ms per token,   726.41 tokens per second)
llama_print_timings:        eval time = 21169.54 ms /   990 runs   (   21.38 ms per token,    46.77 tokens per second)
```

## 7B `q4_k_m`

```plaintext
master
perplexity: 1.13 seconds per pass - ETA 12.32 minutes
llama_print_timings:        load time =   981.64 ms
llama_print_timings:      sample time =    17.23 ms /   500 runs   (    0.03 ms per token, 29019.15 tokens per second)
llama_print_timings: prompt eval time =  1227.96 ms /   385 tokens (    3.19 ms per token,   313.53 tokens per second)
llama_print_timings:        eval time = 13685.93 ms /   495 runs   (   27.65 ms per token,    36.17 tokens per second)

master `LLAMA_CUDA_DMMV_X=64 LLAMA_CUDA_MMV_Y=8 LLAMA_CUDA_KQUANTS_ITER=1`
perplexity: 1.16 seconds per pass - ETA 12.60 minutes
llama_print_timings:        load time =   984.89 ms
llama_print_timings:      sample time =    17.34 ms /   500 runs   (    0.03 ms per token, 28843.38 tokens per second)
llama_print_timings: prompt eval time =  1229.90 ms /   385 tokens (    3.19 ms per token,   313.03 tokens per second)
llama_print_timings:        eval time = 13771.24 ms /   495 runs   (   27.82 ms per token,    35.94 tokens per second)

pull
perplexity: 0.99 seconds per pass - ETA 10.77 minutes
llama_print_timings:        load time =   983.22 ms
llama_print_timings:      sample time =    14.98 ms /   500 runs   (    0.03 ms per token, 33384.52 tokens per second)
llama_print_timings: prompt eval time =  1101.86 ms /   385 tokens (    2.86 ms per token,   349.41 tokens per second)
llama_print_timings:        eval time = 13699.62 ms /   495 runs   (   27.68 ms per token,    36.13 tokens per second)
```

</details>