@JohannesGaessler 2023-08-31
If you use the `llama-bench` binary it will automatically run the benchmark multiple times, average the results, and print a GitHub table.