@Ronny-apo 2023-08-31
My results:

Master:

Q4_K

| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| LLaMA 7B mostly Q4_K - Medium  |   3.94 GiB |     6.74 B | ROCm       |  99 | pp 512     |    602.43 ± 0.40 |
| LLaMA 7B mostly Q4_K - Medium  |   3.94 GiB |     6.74 B | ROCm       |  99 | tg 128     |     42.63 ± 0.02 |

Q4_0

| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | ROCm       |  99 | pp 512     |    747.85 ± 1.61 |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | ROCm       |  99 | tg 128     |     51.35 ± 0.26 |

Master with overrides (LLAMA_CUDA_DMMV_X=64 LLAMA_CUDA_MMV_Y=8 LLAMA_CUDA_KQUANTS_ITER=1):

Q4_K

| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| LLaMA 7B mostly Q4_K - Medium  |   3.94 GiB |     6.74 B | ROCm       |  99 | pp 512     |    602.19 ± 1.54 |
| LLaMA 7B mostly Q4_K - Medium  |   3.94 GiB |     6.74 B | ROCm       |  99 | tg 128     |     42.84 ± 0.10 |

Q4_0

| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | ROCm       |  99 | pp 512     |    748.54 ± 1.30 |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | ROCm       |  99 | tg 128     |     51.69 ± 0.22 |


PR #2910:

Q4_K

| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| LLaMA 7B mostly Q4_K - Medium  |   3.94 GiB |     6.74 B | ROCm       |  99 | pp 512     |    691.66 ± 0.64 |
| LLaMA 7B mostly Q4_K - Medium  |   3.94 GiB |     6.74 B | ROCm       |  99 | tg 128     |     42.61 ± 0.01 |


Q4_0

| model                          |       size |     params | backend    | ngl | test       |              t/s |
| ------------------------------ | ---------: | ---------: | ---------- | --: | ---------- | ---------------: |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | ROCm       |  99 | pp 512     |    722.16 ± 0.42 |
| LLaMA 7B mostly Q4_0           |   3.56 GiB |     6.74 B | ROCm       |  99 | tg 128     |     51.48 ± 0.01 |