@KerfuffleV2 2023-08-31
@Ronny-apo Might want to mention your hardware.