@Engininja2 2023-09-03
I tried finding tunings for the RX 5700XT. I had the best results with `nwarps=8, min_warps_per_eu=8, mmq_x=64, mmq_y=64` except for q5_k and q6_k where `mmq_x=32` was faster, and q_4k where `mmq_x=32` was slightly faster or tied. Not specifying launch_bounds was as fast or slightly faster.

Model | t/s master | t/s PR | t/s RDNA1 | t/s RDNA1 w/o launch_bounds | Speedup
-- | -- | -- | -- | -- | --
LLaMA v2 7B mostly Q2_K | 195.38 | 177.55 | 205.31 | 205.59 | 1.05
LLaMA v2 7B mostly Q3_K - Small | 202.27 | 194.90 | 214.74 | 215.16 | 1.06
LLaMA v2 7B mostly Q4_0 | 306.56 | 297.59 | 306.56 | 307.00 | 1.00
LLaMA v2 7B mostly Q4_1 | 283.73 | 287.28 | 299.70 | 300.12 | 1.06
LLaMA v2 7B mostly Q4_K - Small | 268.10 | 225.29 | 272.24 | 272.59 | 1.02
LLaMA v2 7B mostly Q5_0 | 295.97 | 268.46 | 292.03 | 296.21 | 1.00
LLaMA v2 7B mostly Q5_1 | 294.45 | 275.77 | 288.85 | 294.56 | 1.00
LLaMA v2 7B mostly Q5_K - Small | 263.37 | 217.71 | 273.29 | 273.27 | 1.04
LLaMA v2 7B mostly Q6_K | 251.19 | 218.67 | 253.96 | 254.01 | 1.01
LLaMA v2 7B mostly Q8_0 | 301.39 | 291.15 | 289.27 | 301.46 | 1.00