@sozforex 2023-09-11
On RX 6850M XT (RDNA2, navi22, gfx1031) with gfx1030 override this PR does not affect token generation time (<1% difference), and possibly improves prompt eval time.

model_type | test | t/s PR | t/s master | speedup
-- | -- | --: | --: | --:
LLaMA v2 7B mostly Q2_K | tg 128 | 45.56 | 45.58 | 99.94%
LLaMA v2 7B mostly Q3_K - Small | tg 128 | 42.82 | 43.03 | 99.52%
LLaMA v2 7B mostly Q3_K - Medium | tg 128 | 45.47 | 45.56 | 99.81%
LLaMA v2 7B mostly Q3_K - Large | tg 128 | 44.54 | 44.17 | 100.86%
LLaMA v2 7B mostly Q4_0 | tg 128 | 67.71 | 67.71 | 100.01%
LLaMA v2 7B mostly Q4_K - Small | tg 128 | 51.00 | 51.11 | 99.79%
LLaMA v2 7B mostly Q4_K - Medium | tg 128 | 50.28 | 50.59 | 99.38%
LLaMA v2 7B mostly Q5_0 | tg 128 | 59.10 | 59.04 | 100.09%
LLaMA v2 7B mostly Q5_K - Small | tg 128 | 48.30 | 48.17 | 100.27%
LLaMA v2 7B mostly Q5_K - Medium | tg 128 | 48.08 | 47.98 | 100.22%
LLaMA v2 7B mostly Q6_K | tg 128 | 46.08 | 45.97 | 100.23%
LLaMA v2 7B mostly Q8_0 | tg 128 | 41.97 | 42.23 | 99.40%
LLaMA v2 7B mostly Q2_K | pp 512 | 438.78 | 430.63 | 101.89%
LLaMA v2 7B mostly Q3_K - Small | pp 512 | 436.26 | 396.38 | 110.06%
LLaMA v2 7B mostly Q3_K - Medium | pp 512 | 492.68 | 467.86 | 105.31%
LLaMA v2 7B mostly Q3_K - Large | pp 512 | 506.46 | 433.55 | 116.82%
LLaMA v2 7B mostly Q4_0 | pp 512 | 693.01 | 682.52 | 101.54%
LLaMA v2 7B mostly Q4_K - Small | pp 512 | 603.28 | 635.62 | 94.91%
LLaMA v2 7B mostly Q4_K - Medium | pp 512 | 598.84 | 628.73 | 95.25%
LLaMA v2 7B mostly Q5_0 | pp 512 | 693.96 | 606.20 | 114.48%
LLaMA v2 7B mostly Q5_K - Small | pp 512 | 657.18 | 517.15 | 127.08%
LLaMA v2 7B mostly Q5_K - Medium | pp 512 | 643.89 | 526.45 | 122.31%
LLaMA v2 7B mostly Q6_K | pp 512 | 591.70 | 579.77 | 102.06%
LLaMA v2 7B mostly Q8_0 | pp 512 | 678.58 | 665.16 | 102.02%